Updated
Changed the title of the commands list from the \help chatcommands sub-command from "Chat Commands Syntax" to "Chat Commands List"; the \rtp commands now defaults to not requiring any tags to be used, added \\itfill and \\idtfill commands that are the same as the \itfill and \idtfill commands except that they don't require the corners/radi of the shape to be entered and will instead use the selected area values from pos1 and pos2; the custom Vector class is now available for use in script eval without requiring the use of imports; added the \getworldspawnpoint command with the following aliases: \getworldspawn, \getwsp, \getws, \gwsp, and \gws; the \remexp command now also remove lit tnt and tnt minecarts within the specified radius; added the \remexpne command which is like the old version of the \remexp command where it will only remove explosive block but not entities; added the following brush types: cube, square, splattercube, splattersquare, splattersurface, splattercubesurface, splattersquaresurface, and remexpne. replaced all the new recipe files with updated versions again.