Massive Update.
Added the \\backuparea command.
Added the \snapshot command.
Added the \removeotheritemenchantments command with the following aliases: \remotheritemenchants and \roie.
The BlockPattern class now allows for putting r: or s: before the pattern text to specify wether the pattern is in random mode (default) or sequence mode respectively.
Added the formats for the save, load, copytodisk, copytomemory, and removeair sub-commands of the \structure command to the \help command.
Added the format for the \\protectarea command to the \help command.
Added the l flag to the \butcher and \butcherdespawn commands.
Added the k and t flags to the \\undo command.
Fixed the \top command.
Fixed the save, createempty, load, copy, copytodisk, copytomemory, and savetodisk sub-commands of the \structure command.
Made the list of structures outputted by the \liststructures command be sorted in alphabetical order.
Made the undo clipboard default to being saved to the world files instead of in memory.
Fixed the lore settings for the \brush command.
Added the warden to the list of boss types that area kill/despawned by the \butcher and \butcherdespawn commands.
Added the \aligncenter command.
Fixed the \jumpto command.
Fixed the \\cut and \\copy commands when cutting/pasting an area too large to fit inside of a single structure.
Added a warning for when there is nothing else left to undo.
Changed some of the icons for some of the sub-menus in the main settings menu.
Added the Script Settings, UI Settings, and Manage Game Rules sub-menus to the main settings menu.
Made a lot of the menus go back to the previous menu when the x button is pressed instead of closing the menu altogether.
Added the following options to the settings sub-menus: playerDataRefreshRate and maxPlayersPerManagePlayersPage.
Added 2 new structures to allow for getting all enchantments with levels from -2 to 294 as well as level -32k: andexdb:-2-294_steb and andexdb:-32k_steb.
Internal:
Moved the config class from commands.ts to Main.ts.
Fixed the getGroundSolidBlock() function.
Added the testForObjectExtension() function.
Added the global constant dimensionsb.
Added the global constant dimensionsc.
Added the getEntityHeldItemSlot() function.
Added the getPlayerHeldItemSlot() function.