Updated
Added the \gettransformst command.
Added the formatting and descriptions for the \align and \aligncenter commands to the \help command.
Updated the \help chatcommands sub-command.
Added the info about the l flag for the \butcherdespawn command to the \help command.
Added the \help js alias to the \help jsfunction sub-command.
Updated the syntax information for the \help command.
Added the format for the andexdb:spawnSimulatedPlayer /scriptevent command to the \help scriptevent sub-command.
Fixed the text on the Previous Page and Next Page buttons on the manage players menu.