Fixed the \rtp command.
The \rtp command now correctly uses the name of the player who sent the teleport request instead of the person that the teleport request is targeted at in the feedback to the target of the teleport request.
The \rtp command teleport requests now actually expire after a minute instead of only expiring once the player that it was targeted at actually leaves the world/realm/server.
Updated the feedback messages to the sender of the teleport request for when the teleport request expires or times out.
Changed the command versions of the \tpaccept and \tpdeny commands from 1.0.0 to 0.0.1-alpha and changed their command colors to be red instead of white since they are not functional at all yet.