updated
Added a LICENSE.txt file.
Deprecated a lot of functions.
Added JSDoc documentation to a lot of functions.
Added a few new generation functions.
Fixed an issue where the \snapshot listbackups sub-command would list backups multiple times depending on how many structures were used to save that backup.
Updated the debug info for the entity debug stick.
Updated the actionbar debug info displayed when sneaking while holding the editor stick.
Massive overhaul to the chat ranks system.
There is now an option to show chat ranks on player name tags.
Updated the andexdb:chatCommands scriptevent to allow for escaping < and > with \\<\\ and \\>\\ respectively.
Added the andexdb:cmd scriptevent which is the same as the andexdb:chatCommands scriptevent except for that it runs the command with the the player input being an executeCommandPlayerW class to allow for running commands from a command block directly.
Added the s:eb, is:eb, s:elcb, and is:elcb scriptevents.
Added the andexdb:xz_axis_particle_pos1 particle.
Changed the formatting code for the \\generatejs command from §f to §c.
Added the \execute command.
Major update to the executeCommandPlayerW class.
Added the parseBlockMatcherType() function.
Added the "fromExecute?: boolean" property to the params parameter of the chatCommands function.
The chatCommands function now assigns the player property with an (executeCommandPlayerW & {readonly id: string, readonly typeId: string, readonly name: string}) instance instead of merging the executeCommandPlayerW instance with the player and then assigning the player property with the player.
When the chatCommands function is run by an entity, the server notification message will now show the nameTag of the entity or the typeId of the entity if the nameTag is empty, followed by the id of the entity inside of arrow brackets (<>).
Added error catching to the \morph, \tint, \scale, \visualscale, \visualscaleenabled, and \managescriptautoeval commands.
Added the \editorstick, \editorstickb, \editorstickc, and \notificationsettings commands.
Added the \\shift and \\offset commands.
Updated tyhe \\replace command to use a blockReplacer parameter type for the block mask instead of a string parameter and a blockStates parameter.
Added the \\set, \\seti, \\flood, \\remove, \\walls, \\sphere, \\hsphere, and \\cone commands.
Added the chat ranks settings section to the settings menu.
Added the evalBeforeEvents:chatSendBeforeModifiedMessageEval dynamic property.
Fixed the sphere generation to not do the repeat generation of each block in the z axis for each block in the z axis.