Added Release Files And Updated
Added release files.
The \rtp command and all references to it have now been changed to \tpa.
Fixed the target selector parameters for when the target selector is a player's name without any quotation marks.
Removed an accidentally left in console.warn() for when a custom chat message is sent.