Updated
Fixed the ban system for players with spaces in their names.
Fixed the \\remove, \\set, \\seti, and \\flood commands.
Fixed an issue where the block mask parameters of a lot of commands would require the block type to be specified with the namespace, so minecraft:air would work but air would not work.