Updated
Added release files for 1.18.1.
Fixed the random and shuffled gradients.
Added the cycleRGB async function.
When a command is run by another player though the \execute command, it will now execute regardless of the targeted player's permission levels.
Fixed the \kick command for player's with spaces in their names.
Fixed an issue where the hour 12 would display as 00 and the hour 11 would display as -1.
Fixed the ban system for players with spaces in their names.