Updated
Added the undoClipboardMode settings option to the script settings section.
Added the PlayerNotifications class.
Added a bunch of new options to the Notifications Settings section. Including the ability to add sounds to the notifications.
Updated the entity debug stick debug log info.
Updated the home system format version from 0.0.1-indev.1 to 0.7.0-beta.72.
Added the \spawn command and the spawnCommandLocation settings option.
Added the SemVerString class.
Added the cullNull, cullUndefined, and cullEmpty functions.
Updated the \help cmd sub-command.
Added the following sub-commands to the \help command: cmdextra, cmddebug, and customcmddbug.
Fixed an issue with the flags for the \\cut, \\copy, and \\paste commands not showing up in the \help cmd sub-command.
Added the \setplayernametag command.
Fixed an issue where many WorldEdit commands would require you to set a block mask in order to work.