Updated
Changed the parameter type of the \\replace command from SingleBlockMask to mask.
Added fully functional axis markings to appear on a block for 2 seconds when you select it with the selection tool.
Fixed a typo in the custom settings section for the add-on where the word must was spelled mmust instead of must.
Updated the custom JSONParse and JSONStringify functions to support the BigInt types.
Added the \attribute and \data commands, but they are still in early development.
The \help command now tells you when a command is deprecated or non-functional.
Fixed the Global Settings section of the main menu.
Added a bunch of new options to the notifications settings section.
Undo clipboard items that were only saved to memory will now be deleted upon a world restart.
Removed a bunch of unnecessary quotation marks from the commands_list file.
Added descriptions for the \\drain, \\itfill, and \\idtfill commands to the \help command.
Updated the syntax that the \help command displays for a few WorldEdit commands.
Added a new section to the \help chatcommands sub-command for the deprecated commands.
If a command is marked as hidden in the commands_list file, it will now not show up in the \help commands, and only show up in the new \help cmddebugplus sub-command.
Added the \help cmddebugplus sub-command.
Fixed a bug where BlockPatterns would not allow for block states to be specified using square brackets (ex. ["my_state"="value","on"=true,"time"=1234]).
Added the parseSNBT() function.
Made a massive performance improvement to all of the commands.
Added the \item keepondeath, \item lockmode, \item slot <slot> keepondeath, and \item slot <slot> lockmode sub-commands to the \item command.
Fixed the \getuuid command.
Added the -h flag to the \\paste command to make the copied structure be pasted at your current location instead of the selected location.
Fixed the editor stick information for the jukeboxes.
Fixed the entity controller's displayed information.