Added Preview Release Files And Updated
Deprecated the "Inventory Transfer" section of the main menu.
The gametestStructureDefaultSpawnLocation settings option now is formatted as "x y z" instead of "x, y, z".
Fixed the spawnCommandLocation settings option.
Fixed a few typos in the description of the undoClipboardMode settings option.
Added the protectedAreasRefreshRate option to the script settings section.
Finished the notification settings section.
Massive overhaul to the Anti-Spam System, now the anti-spam will only activate if the player sends a number of messages equal to (Message count to trigger anti-spam) and those messages each had a delay of at most (Maximum time between messages) seconds between them. Also, added an option to make it so that when a player tried sending a message while they are muted by the anti-spam system timer, then the timer will restart again, this is to prevent people who are using a bot to spam from being only muted for a certain amount of time, and then sending a few more messages before being muted again.
Added the doNotSetNameTag option to the personal settings section.
Deprecated the "Eval Auto Script/Execute Settings" settings section.
Fixed a bug where when an explosion occured, a block was broken by a player, a block was interacted with by a player, a block was placed by a player, or an entity was interacted with by a player, and there were no restricted/protected areas of that type set, then it would cause errors in the content error log to appear.
Fixed the version numbers and formatting codes of a bunch of commands.
The \help cmd and \help cmdextra sub-commands will now show the syntax or description listed in the definition of the command, if it can not find it in the commandsyntaxes or commanddescriptions object.
When running script eval through the chat with ${se} or ${scripteval}, it will now once again execute it from the commands file. You can still run the script eval from inside the chatMessage function by using ${sel} or ${scriptevallocal} instead.
Added the global tempVariables object.
Added the cmdslist and cmdsdocs imports to the Main file.
Added maxTimeBewteenMessagesToTriggerAntiSpam and protectedAreasRefreshRate to config.
Added the andexdb:silentCmd /scriptevent command to run custom commands without notifying players in the chat, this is very usefull for running custom commands inside of command blocks.
Added the andexdb:silentBuiltInCmd, andexdb:builtInCmd, andexdb:silentCustomCmd, and andexdb:customCmd /scriptevent commands to allow for pre-defining whether a command is built-in or custom, to increase performance when running custom commands with a command block.
Added a lot more to the in-game documentation in the how-to-play screen.
Updated the LICENSE.txt files for the preview versions.