Updated
Added/Updated the syntax and descriptions displayed in the \help command for many commands.
Fixed a bug where commandSettings class instances would have the wrong number of backslashes in the command names of the WorldEdit commands.
The getEntityHurtNotifications notifications now also display the velocity of the hurt entity.
Fixed the \br, \\br, and \\brush aliases of the \brush command (ANDEXDB-40)
Fixed a bug where RemExp brushes and the \rexemp command would fail to remove explosive entities (ANDEXDB-44)
Fixed a typo with the following commands and sub-commands where they would say "Seccessfully" instead of "Successfully":
\structure saveempty
\structure createempty
\structure delete
\structure load
\structure copy
\structure copytodisk
\structure copytomemory
\structure savetodisk
\structure movetomemory
\timezone
\\pos1
\\pos2
\\hpos1
\\hpos2
\\selectmode
\\chunk
\\shift
\brush (ANDEXDB-42)
Fixed a bug where the \home command would say you have too many homes if you try and create a home with the same name as an already existing home (ANDEXDB-38)