Updated
Changes:
World Border System:
The world border system is now functional in all dimensions.
Added a bunch of new options to the world border system.
The world border dust particles are now visible when the shaders are enabled.
Items:
The ultra bundle item is no longer hand equipped.
Fixes:
Commands:
Fixed the \execute in sub-command when using the \execute command to run the \run command.
The \run command now sends the success number to the source of the \execute command instead of the entity that the \execute command targeted, if the \run command was run using the \execute command.