Updated
Changed how some of the menus work so that they are async functions that will only go to the menu they were opened from, or if they were not opened from a async function that was made for this, it will just close the whole ui instead of going back.
Updated the player save start and stop function.
Added the advanced settings section to the settings menu (Main Menu>Settings>Advanced Settings).
Added the debug menu to the advanced setting menu to allow stopping certain services such as the banned players checker, protected areas refresher, and the player data save system (Main Menu>Settings>Advanced Settings>Debug).
Added the `\pthru` command, it is the same as \thru except it can also teleport you into 1 block tall spaces.
Added the subscription ids for a lot more of the subscribed events to the global `subscribedEvents` property.